requirements.txt
streamlit
streamlit-audiorecorder
pydub
python-dotenv
groq
ollama